Gay shlafen:  Yiddish for "go to sleep".
Now doesn't "gay shlafen" have a softer, more soothing sound than the
harsh, staccato "go to sleep"?  Listen to the difference:
	"Go to sleep, you little wretch!" ... "Gay shlafen, darling."
Obvious, isn't it?
	Clearly the best thing you can do for you children is to start
speaking Yiddish right now and never speak another word of English as
long as you live.  This will, of course, entail teaching Yiddish to all
your friends, business associates, the people at the supermarket, and
so on, but that's just the point.  It has to start with committed
individuals and then grow....
	Some minor adjustments will have to be made, of course: those
signs written in what look like Yiddish letters won't be funny when
everything is written in Yiddish.  And we'll have to start driving on
the left side of the road so we won't be reading the street signs
backwards.  But is that too high a price to pay for world peace?
I think not, my friend, I think not.
		-- Arthur Naiman, "Every Goy's Guide to Yiddish"